Test the 'reached end of input with reading anything' condition.
This condition fires if the end of a bzip2 compressed structure has been
read, but eof wasn't detected (i.e. the compressed data length was an
exact multiple of READ_BUFFER_SIZE). An additional read is performed to
find a second structure, but reaches eof instead. The position isn't
advanced from the start of the structure so the read is terminated.